fix changes

+++ src/services/authService.js
@@ -1,14 +1,11 @@
 import bcrypt from "bcryptjs";
 import jwt from "jsonwebtoken";
-import RefreshRepository from "../repositories/RefreshRepository";
-import userRepository from "../repositories/userRepository";
+import RefreshRepository from "../repositories/RefreshRepository.js";
+import UserRepository from "../repositories/userRepository.js";
 import validator from "validator";
 import AppError from "../error/AppError.js";
 import { generateToken } from "../utils/refreshUtils.js";
-import { config } from "dotenv";
-import generalValidator from "../utils/userValidator.js";
-
-config();
+import verificationService from "./verificationService.js";
 
 const authService = {
   async login(body) {
@@ -35,7 +32,7 @@ const authService = {
     });
 
     const refreshToken = await RefreshRepository.show({ userInfo: user._id });
-    await refreshToken.deleteOne();
+    if (refreshToken) await refreshToken.deleteOne();
 
     const rawToken = generateToken();
     const hashedToken = await bcrypt.hash(rawToken, 10);
@@ -43,6 +40,7 @@ const authService = {
     const newRefresh = await RefreshRepository.create({
       userInfo: user._id,
       token: hashedToken,
+      expiresAt: new Date(Date.now() + 7 * 24 * 60 * 60 * 1000),
     });
 
     await newRefresh.populate({
@@ -51,6 +49,69 @@ const authService = {
     });
 
     return { user, rawToken, acessToken };
+  },
+
+  async activateAccount(body, token) {
+    const { success } = verificationService.compareCode(token?.code, body.code);
+
+    if (success) {
+      const activeUser = await UserRepository.update(
+        { _id: token.id },
+        { isActive: true },
+      );
+
+      const acessToken = jwt.sign(
+        { id: activeUser._id, email: activeUser.email },
+        process.env.JWT_SECRET,
+        {
+          expiresIn: "15m",
+        },
+      );
+
+      const rawToken = generateToken();
+      const hashedToken = await bcrypt.hash(rawToken, 10);
+
+      const refreshToken = await RefreshRepository.create({
+        token: hashedToken,
+        userInfo: activeUser._id,
+        expiresAt: new Date(Date.now() + 7 * 24 * 60 * 60 * 1000),
+      });
+
+      await refreshToken.populate({
+        path: "userInfo",
+        select: "_id email",
+      });
+
+      return { activeUser, acessToken, refreshToken };
+    }
+
+    throw new AppError(`Código errado.`, 400);
+  },
+
+  async resendCode(body) {
+    const { email } = body;
+
+    if (!email || !validator.isEmail(email)) {
+      throw new AppError(`E-mail inválido`, 400);
+    }
+
+    const user = await UserRepository.show({ email });
+
+    if (!user) throw new AppError(`Usuário não encontrado`, 404);
+
+    const { code } = await emailService.sendVerificationMail(email);
+
+    const payload = {
+      id: user._id,
+      code,
+      type: "ativacao",
+    };
+
+    const token = jwt.sign(payload, process.env.JWT_SECRET, {
+      expiresIn: "6m",
+    });
+
+    return { success: true, token };
   },
 
   async refresh(token) {
